Update RNN_Seattle.py
Removed the training from this code. Before there was an if statement that was checking to see if there was a trained model and if there was not then it would train a model and save it. The training of the model is being moved to train.py This code now only makes the prediction and saves those predictions